requirements.txt
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
streamlit==1.35.0
schedule==1.2.2
pytz==2024.1
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
asyncpg==0.29.0
plotly==6.7.0
yfinance==0.2.40
flask==3.0.3
